arch-auth/pyproject.toml
[build-system]
requires = ["setuptools>=75.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auth-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "bcrypt==4.0.1",
    "asyncpg==0.29.0",
    "cffi==1.17.1",
    "click==8.1.8",
    "cryptography==44.0.0",
    "dnspython==2.7.0",
    "email_validator==2.1.2",
    "fastapi==0.115.6",
    "h11==0.14.0",
    "idna==3.10",
    "lazy-model==0.2.0",
    "makefun==1.15.6",
    "packaging==24.2",
    "sqlalchemy[asyncio]==2.0.25",
    "passlib[bcrypt]==1.7.4",
    "pycparser==2.22",
    "pydantic==2.10.5",
    "pydantic_settings==2.9.1",
    "pydantic_core==2.27.2",
    "psycopg2-binary==2.9.11",
    "PyJWT==2.8.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.7",
    "setuptools==75.8.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.41.3",
    "toml==0.10.2",
    "typing_extensions==4.12.2",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
]

[project.optional-dependencies]
test = [
    "asgi-lifespan==1.0.1",
    "httpx==0.28.1",
    "pytest==9.0.2",
    "pytest-asyncio==0.18.3",
    "pytest-cov==3.0.0",
    "pytest-mock==3.14.0",
    "aiosqlite==0.22.1"
]

[tool.setuptools.packages.find]
include = ["api*"]

[tool.coverage.html]
directory = "tests/cov_html"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = [
    "--cov=api",
    "--cov-report=html",
    "--tb=line",
    "-s"]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
    "login: tests for user login"
]
